requirements.txt
yt-dlp>=2024.10.7
openai-whisper>=20240930
